// src/app/model/wallet-by-userid.model.ts
export class WalletByUserId{
    walletId: number = 0;
    walletCode: string = "";
    walletName: string = "";
    userId: number = 0;
    totalAmount: string = "";
    availableAmount: string = "";
    blockedAmount: string = "";
    currencyId: number = 0;
    symbolsDisplay: string = "";
    isDefault: string = "";
    isAdmin: string = "";
}