// src/app/model/transaction-manager.model.ts
export class TransactionManager{
    walletTransactionsId: number = 0;
    walletTransactionsCode: string = "";
    walletId: number = 0;
    amount: string = "";
    balance: string = "";
    walletTransactionsType: string = "";
    walletTransactionsTypeName: string = "";
    content: string = "";
    transactionsDate: string = "";
    receiptId: number = 0;
    status: string = "";
    statusDisplay: string = "";
    currencyId: string = "";
    symbolsDisplay: string = "";
    symbolsLocation: string = "";
    exchangeRate: string = "";

    public TransactionManager(){};
}